// geidea-pilgrim-wallet-test/src/api/errorTypes.ts
export enum ErrorType {
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  INVALID_INPUT_FORMAT = "INVALID_INPUT_FORMAT",
  UNVERIFIED_PHONE_NUMBER = "UNVERIFIED_PHONE_NUMBER",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  FRAUDULENT_USER = "FRAUDULENT_USER",
  UNKNOWN_SEON_ERROR = "UNKNOWN_SEON_ERROR",
  INCORRECT_CODE = "INCORRECT_CODE",
  VERIFICATION_NOT_FOUND = "VERIFICATION_NOT_FOUND",
  EMAIL_DOES_NOT_EXISTS = "EMAIL_DOES_NOT_EXISTS",
  PHONE_NUMBER_ALREADY_EXISTS = "PHONE_NUMBER_ALREADY_EXISTS",
  UNKNOWN_TWILIO_ERROR = "UNKNOWN_TWILIO_ERROR",
  PHONE_NUMBER_DOES_NOT_EXISTS = "PHONE_NUMBER_ALREADY_VERIFIED",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  USERS_ALREADY_IN_DATABASE = "USERS_ALREADY_IN_DATABASE",
  BAD_REQUEST = "BAD_REQUEST",
  NOT_FOUND = "NOT_FOUND",
  OWN_ACCOUNT_CONTACT = "OWN_ACCOUNT_CONTACT",
  NICKNAME_ALREADY_USED = "NICKNAME_ALREADY_USED",
  CONTACT_ALREADY_EXISTS = "CONTACT_ALREADY_EXISTS",
  CONTACT_NOT_FOUND = "CONTACT_NOT_FOUND",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  INVALID_EXPO_TOKEN = "INVALID_EXPO_TOKEN",
  INVALID_PHONE_NUMBER = "INVALID_PHONE_NUMBER",
  TRANSACTION_NOT_FOUND = "TRANSACTION_NOT_FOUND",
  SUBACCOUNT_NOT_FOUND = "SUBACCOUNT_NOT_FOUND",
  INSUFFICIENT_FUNDS = "INSUFFICIENT_FUNDS",
  UNKNOWN_WALLET_ERROR = "UNKNOWN_WALLET_ERROR",
  WALLET_NOT_AVAILABLE = "WALLET_NOT_AVAILABLE",
  EXTERNAL_SEND_INSIDE_THE_SYSTEM = "EXTERNAL_SEND_INSIDE_THE_SYSTEM",
  INVALID_AMOUNT_NEGATIVE_OR_ZERO = "INVALID_AMOUNT_NEGATIVE_OR_ZERO",
  INVALID_AMOUNT_TOO_MANY_DECIMAL_PLACES = "INVALID_AMOUNT_TOO_MANY_DECIMAL_PLACES",
  INVALID_BANK_ACCOUNT = "INVALID_BANK_ACCOUNT",
  BANK_ACCOUNT_ALREDY_ADDED = "BANK_ACCOUNT_ALREDY_ADDED",
  DUPLICATE_BANK_ACCOUNT = "DUPLICATE_BANK_ACCOUNT",
  USER_BANK_ACCOUNT_NOT_FOUND = "USER_BANK_ACCOUNT_NOT_FOUND",
  CANNOT_ALLOW_LESS_THEN_MIN_FIAT = "CANNOT_ALLOW_LESS_THEN_MIN_FIAT",
  CANNOT_ALLOW_MORE_THEN_MAX_FIAT = "CANNOT_ALLOW_MORE_THEN_MAX_FIAT",
  UNSUPPORTED_TOKEN = "UNSUPPORTED_TOKEN",
  ISSUANCE_BANK_NOT_FOUND = "ISSUANCE_BANK_NOT_FOUND",
  ISSUANCE_WALLET_USER_DISABLE = "ISSUANCE_WALLET_USER_DISABLE",
  WALLET_DISABLED = "WALLET_DISABLED",
  TRANSACTION_IS_ALREADY_PAID = "TRANSACTION_IS_ALREADY_PAID",
  SEND_PHONE_OTP_COMMON = "SEND_PHONE_OTP_COMMON",
  VERIFY_PHONE_OTP_COMMON = "VERIFY_PHONE_OTP_COMMON",
  SALT_KEY_NOT_FOUND = "SALT_KEY_NOT_FOUND",
  INVALID_QR = "INVALID_QR",
  BLOCKCHAIN_SERVICE_NOT_FOUND = "BLOCKCHAIN_SERVICE_NOT_FOUND",
  WALLET_DISABLED_RECIPIENT = "WALLET_DISABLED_RECIPIENT",
  BLOCKCHAIN_INSUFFICIENT_AMOUNT = "BLOCKCHAIN_INSUFFICIENT_AMOUNT",
  RECIPIENT_WALLET_NOT_ACTIVATED= "RECIPIENT_WALLET_NOT_ACTIVATED",
  OTHER_BANK_TRANSFER_NOT_ALLOWED = "OTHER_BANK_TRANSFER_NOT_ALLOWED"
}